centralise toggle switches
